+++ hilt/annotations/src/main/kotlin/FactoryMethod.kt
@@ -32,6 +32,7 @@ import kotlin.reflect.KClass
  */
 @Target(AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.BINARY)
+@Repeatable
 @MustBeDocumented
 annotation class FactoryMethod(