+++ hilt/annotations/src/main/kotlin/TestBoundTo.kt
@@ -29,6 +29,7 @@ import kotlin.reflect.KClass
  */
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.BINARY)
+@Repeatable
 @MustBeDocumented
 annotation class TestBoundTo(