+++ hilt/processor-ksp/src/main/kotlin/model/Model.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020-2023 Slawomir Czerwinski
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package it.czerwinski.android.hilt.processor.model
+
+import com.google.devtools.ksp.symbol.KSFile
+import com.squareup.kotlinpoet.AnnotationSpec
+import com.squareup.kotlinpoet.ClassName
+
+interface Model {
+    val declarationFile: KSFile
+    val moduleKind: HiltModuleKind
+    val moduleClassName: ClassName
+    val replacesModuleClassName: ClassName
+    val componentClassName: ClassName
+    val annotations: List<AnnotationSpec>
+    val isTest: Boolean
+}